Moved the scraper from the root folder to its own folder under the app folder
