+++ .env.example
@@ -1,0 +1,1 @@
+MONGODB_URI="mongodb+srv://<username>:<password>@<cluster_url>/?retryWrites=true&w=majority&appName=<cluster_name>"
