+++ course_extractor/app/storage/mongodb_storage.py
@@ -1,8 +1,10 @@
-from AutoRevise.course_extractor.app.db.connector import get_mongodb_connection
-from AutoRevise.course_extractor.app.db.models import CourseDocument
+from bson import ObjectId
+
+from course_extractor.app.db.connector import get_mongodb_connection
+from course_extractor.app.db.models import CourseDocument
 
 
-def save_course_to_mongodb(course_document):
+def save_course_to_mongodb(course_document: CourseDocument, collection: str = "python_courses") -> ObjectId:
     """
     Save a CourseDocument object to the `python_courses` collection in the `courses_db` database.
 
@@ -11,12 +13,13 @@ def save_course_to_mongodb(course_document):
 
     Returns:
         ObjectId: The ID of the inserted document.
+        :param collection: Name of the target database collection.
     """
     # Get the MongoDB connection
     db = get_mongodb_connection()
 
-    # Access the `python_courses` collection in the `courses_db` database
-    collection = db["python_courses"]
+    # Access the given collection (default is python_courses) in the `courses_db` database
+    collection = db[collection]
 
     # Insert the document and return the inserted ID
     result = collection.insert_one(course_document.to_dict())
