+++ course_extractor/app/utils/logger.py
@@ -1,5 +1,14 @@
 import logging
+from datetime import datetime
+
 
 def setup_logger():
     logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s")
     return logging.getLogger(__name__)
+
+
+def log_message(message: str):
+    """Prints the current time with a given message"""
+
+    current_time = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+    print(f"{current_time} | {message}")
