--- scraper/.env.example
@@ -1,4 +1,0 @@
-DOWNLOAD_PATH="/download/path/"
-MONGODB_ATLAS_URI="mongodb+srv://<username>:<password>@<cluster_url>/?retryWrites=true&w=majority&appName=<cluster_name>"
-MONGODB_DB_NAME="db_name"
-MONGODB_COLLECTION_NAME="collection_name"
